fix a float->enum conversion error with g++ on big endian systems
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5747 d7cf8633-e32d-0410-b094-e92efae38249

+++ mvm_cmds.c
@@ -654,7 +654,7 @@ void VM_M_setserverlistsort( void )
 {
 	VM_SAFEPARMCOUNT( 2, VM_M_setserverlistsort );
 
-	serverlist_sortbyfield = (serverlist_infofield_t) PRVM_G_FLOAT( OFS_PARM0 );
+	serverlist_sortbyfield = (serverlist_infofield_t)((int)PRVM_G_FLOAT( OFS_PARM0 ));
 	serverlist_sortdescending = (qboolean) PRVM_G_FLOAT( OFS_PARM1 );
 }
